pyproject.toml
[tool.poetry]
name = "portfolio-analyzer"
version = "0.1.1"
description = ""
authors = ["Tapan Pandita <tapan.pandita@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
yfinance = "^0.1.63"
scipy = "^1.7.1"
python-dateutil = "^2.8.2"
tabulate = "^0.8.9"
click = "^8.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
portfolio = "portfolio_analyzer.cli:main"